refactor(ablate): arm parsing and report assembly move out; run_ablation decomposed under the structural gate
The second exact-head review's P2-B: ablate.py had grown past the
1,000-line gate (1,036) and run_ablation() measured C901 14 / 16
branches / 63 statements. Behavior-neutral split, no test edits:

* benchflow.ablate_arms (new, 481 lines) — the pre-flight half:
  everything decidable from the request alone. Arm parsing
  (parse_arm/parse_arms/_split_arm_specs, the arm-kind constants and
  AblationArm), stage/arm validation (validate_arms_for_stage,
  CAPTURABLE_STAGES), the fail-closed task and environment-binding
  resolvers, and the ablation error hierarchy beside its earliest
  raisers. benchflow.ablate re-exports every moved name, so external
  callers and tests keep their import paths.
* benchflow.ablate (674 lines) — orchestration only. run_ablation()
  is decomposed into named phase functions with the same
  failure-isolation semantics, one per leg: _preflight_environment_arms
  (env-arm content gates + stamps, before the parent run costs
  anything), _run_parent (unchanged), _branch_into_arms (the
  request-global and research-end-marker skip gates, branch errors
  returned not raised), _retain_and_cleanup (retention strictly before
  the cleanup that destroys the image; cleanup always runs),
  _finalize_report (outcomes, attribution, failure-isolated result
  materialization). run_ablation() now measures C901 2 / 2 branches /
  19 statements.
* Report assembly already lived in benchflow.branch_report; its
  docstring and TYPE_CHECKING import now point at ablate_arms as
  AblationArm's definition site.

Full suite identical before/after (targeted set 441 passed 4 skipped
on both sides; full 6190 passed).
